physics tweaks
git-svn-id: svn+ssh://repo/usr/local/svn/trunk/game@2023 75e209cf-7bc6-457b-8506-78d88e7090b0

+++ src/player.h
@@ -10,7 +10,7 @@ class Player : public Actor {
 
       float z_rotation;
       int thrusting;
-      int shooting;
+      float shooting;
 
       GLuint dl_cockpit;
       GLuint dl_left_wing;
